requirements.txt
anyio==3.6.2
certifi==2022.12.7
cffi==1.15.1
h11==0.14.0
h3==3.7.6
httpcore==0.16.3
httpx==0.23.3
idna==3.4
numpy==1.24.2
pycparser==2.21
python-telegram-bot==20.2
rfc3986==1.5.0
sniffio==1.3.0
timezonefinder==6.1.10
